@@ -1,17 +1,34 @@
 source 'https://rubygems.org'
 
-# TODO: We pretty much just use active record and active mailer, do we need rails?
-gem "mysql2", "~> 0.4.10"
-gem "rails", "~> 4.1"
+gem 'bootsnap'
+gem 'mysql2'
 
-gem "bunny"
-gem "daemons"
-gem "hashie", "~> 3.3"
-gem "migration_comments"
+# We hardly use any of the features of Rails, so lets just load what we need
+# Initially installed with
+# gem 'rails', '~> 6.0'
+# And then extracted the dependencies from the Gemfile.lock
+# I've removed those we don't need. If reinstating a dependency try and match
+# the version to those already included.
+# A full list of rails components can be found in the Gemfile:
+# https://github.com/rails/rails/blob/master/rails.gemspec
+#
+# Be aware when adding in new components that they may require an initializer
+# or additional configuration options.
+# RAILS DEPENDENCIES
+gem 'activemodel', '~> 6.0.3'
+gem 'activerecord', '~> 6.0.3'
+gem 'activesupport', '~> 6.0.3'
+gem 'bundler', '>= 1.3.0'
+gem 'railties', '~> 6.0.3'
+# # RAILS DEPENDENCIES
+
+gem 'bunny'
+gem 'daemons'
+gem 'hashie', '~> 4.0'
+gem 'migration_comments'
 
 group :test, :development do
-  gem 'factory_girl_rails'
-  gem 'mocktra', '~> 1.0.2'
+  gem 'factory_bot_rails'
   gem 'pry'
   gem 'rspec-rails'
 end
@@ -20,6 +37,7 @@ group :development do
   gem 'guard'
   gem 'guard-bundler'
   gem 'guard-rspec'
-  gem 'rubocop'
+  gem 'rubocop', require: false
+  gem 'rubocop-performance', require: false
+  gem 'rubocop-rails', require: false
 end
-
